fix(cli): address PR #121 review comments
- Warn on SnippetTextEdit instead of silently dropping; share isTextEdit
  predicate between applyWorkspaceEdit and planWorkspaceEdit so dry-run
  editCounts match apply behaviour
- Bump engines.node in @lsproxy/cli to >=22.12.0 (commander@15 requirement)
- Fix codegen: integer-based open enum fallback now emits z.number().int()
  instead of z.number() (fixes ErrorCodesSchema / LSPErrorCodesSchema);
  re-run generate:protocol to apply

+++ packages/core/src/protocol/schemas.ts
@@ -59,14 +59,14 @@ export const ErrorCodesSchema = z.union([
   z.literal(-32603),
   z.literal(-32002),
   z.literal(-32001),
-  z.number()
+  z.number().int()
 ]);
 export const LSPErrorCodesSchema = z.union([
   z.literal(-32803),
   z.literal(-32802),
   z.literal(-32801),
   z.literal(-32800),
-  z.number()
+  z.number().int()
 ]);
 export const FoldingRangeKindSchema = z.union([
   z.literal('comment'),
